Introduce agentxReconnectInterval, for fast retry
In a system with a lot of load, it may be reasonable for the
agentxPingInterval to be high - you don't want too much
ping overhead.  However, even if you have a high request
load, you may want to have an aggressive timer for subagent
reconnection, since if you're not connected all you're doing
is failing to receive requests.  Use separate configuration
values for the ping and reconnect intervals (but default
the reconnect interval to "same as ping interval",
to retain the previous behavior.)